+++ lib/vagrant/config/vm.rb
@@ -4,8 +4,6 @@ require 'vagrant/config/vm/provisioner'
 module Vagrant
   module Config
     class VMConfig < Base
-      include Util::StackedProcRunner
-
       attr_accessor :name
       attr_accessor :auto_port_range
       attr_accessor :box
@@ -17,6 +15,7 @@ module Vagrant
       attr_reader :shared_folders
       attr_reader :network_options
       attr_reader :provisioners
+      attr_reader :customizations
       attr_accessor :guest
 
       def initialize
@@ -24,6 +23,7 @@ module Vagrant
         @shared_folders = {}
         @network_options = []
         @provisioners = []
+        @customizations = []
       end
 
       def forward_port(name, guestport, hostport, options=nil)
@@ -64,8 +64,25 @@ module Vagrant
         @provisioners << Provisioner.new(name, options, &block)
       end
 
-      def customize(&block)
-        push_proc(&block)
+      # TODO: This argument should not be `nil` in the future.
+      # It is only defaulted to nil so that the deprecation error
+      # can be properly shown.
+      def customize(command=nil)
+        if block_given?
+          raise Errors::DeprecationError, :message => <<-MESSAGE
+`config.vm.customize` now takes an array of arguments to send to
+`VBoxManage` instead of having a block which gets a virtual machine
+object. Example of the new usage:
+
+    config.vm.customize ["modifyvm", :id, "--memory", "1024"]
+
+The above will run `VBoxManage modifyvm 1234 --memory 1024` where
+"1234" is the ID of your current virtual machine. Anything you could
+do before is certainly still possible with `VBoxManage` as well.
+          MESSAGE
+        end
+
+        @customizations << command if command
       end
 
       def defined_vms